Develop

See merge request maarch/MaarchCourrier!259

+++ src/app/attachment/controllers/AttachmentTypeController.php
@@ -14,6 +14,7 @@
 
 namespace Attachment\controllers;
 
+use Attachment\models\AttachmentModel;
 use Attachment\models\AttachmentTypeModel;
 use Group\controllers\PrivilegeController;
 use Respect\Validation\Validator;
@@ -24,9 +25,48 @@ class AttachmentTypeController
 {
     public function get(Request $request, Response $response)
     {
-        $attachmentsTypes = AttachmentTypeModel::get(['select' => ['*']]);
+        $rawAttachmentsTypes = AttachmentTypeModel::get(['select' => ['*']]);
+
+        $attachmentsTypes = [];
+        foreach ($rawAttachmentsTypes as $rawAttachmentsType) {
+            $attachmentsTypes[$rawAttachmentsType['type_id']] = [
+                'id'                => $rawAttachmentsType['id'],
+                'typeId'            => $rawAttachmentsType['type_id'],
+                'label'             => $rawAttachmentsType['label'],
+                'visible'           => $rawAttachmentsType['visible'],
+                'emailLink'         => $rawAttachmentsType['email_link'],
+                'signable'          => $rawAttachmentsType['signable'],
+                'icon'              => $rawAttachmentsType['icon'],
+                'chrono'            => $rawAttachmentsType['chrono'],
+                'versionEnabled'    => $rawAttachmentsType['version_enabled'],
+                'newVersionDefault' => $rawAttachmentsType['new_version_default']
+            ];
+        }
 
         return $response->withJson(['attachmentsTypes' => $attachmentsTypes]);
+    }
+
+    public function getBydId(Request $request, Response $response, array $args)
+    {
+        $attachmentType = AttachmentTypeModel::getById(['select' => ['*'], 'id' => $args['id']]);
+        if (empty($attachmentType)) {
+            return $response->withStatus(400)->withJson(['errors' => 'Attachment type does not exist']);
+        }
+
+        $attachmentType = [
+            'id'                => $attachmentType['id'],
+            'typeId'            => $attachmentType['type_id'],
+            'label'             => $attachmentType['label'],
+            'visible'           => $attachmentType['visible'],
+            'emailLink'         => $attachmentType['email_link'],
+            'signable'          => $attachmentType['signable'],
+            'icon'              => $attachmentType['icon'],
+            'chrono'            => $attachmentType['chrono'],
+            'versionEnabled'    => $attachmentType['version_enabled'],
+            'newVersionDefault' => $attachmentType['new_version_default']
+        ];
+
+        return $response->withJson($attachmentType);
     }
 
     public function create(Request $request, Response $response)
@@ -46,11 +86,12 @@ class AttachmentTypeController
         }
 
         $id = AttachmentTypeModel::create([
-            'type_id'               => $body['type_id'],
+            'type_id'               => $body['typeId'],
             'label'                 => $body['label'],
             'visible'               => empty($body['visible']) ? 'false' : 'true',
             'email_link'            => empty($body['emailLink']) ? 'false' : 'true',
             'signable'              => empty($body['signable']) ? 'false' : 'true',
+            'chrono'                => empty($body['chrono']) ? 'false' : 'true',
             'icon'                  => $body['icon'] ?? null,
             'version_enabled'       => empty($body['versionEnabled']) ? 'false' : 'true',
             'new_version_default'   => empty($body['newVersionDefault']) ? 'false' : 'true'
@@ -73,7 +114,7 @@ class AttachmentTypeController
             return $response->withStatus(400)->withJson(['errors' => 'Body label is empty or not a string']);
         }
 
-        $attachmentType = AttachmentTypeModel::getById(['select' => 1, 'id' => $args['id']]);
+        $attachmentType = AttachmentTypeModel::getById(['select' => [1], 'id' => $args['id']]);
         if (empty($attachmentType)) {
             return $response->withStatus(400)->withJson(['errors' => 'Attachment type does not exist']);
         }
@@ -87,6 +128,9 @@ class AttachmentTypeController
         }
         if (isset($body['signable'])) {
             $set['signable'] = empty($body['signable']) ? 'false' : 'true';
+        }
+        if (isset($body['chrono'])) {
+            $set['chrono'] = empty($body['chrono']) ? 'false' : 'true';
         }
         if (isset($body['versionEnabled'])) {
             $set['version_enabled'] = empty($body['versionEnabled']) ? 'false' : 'true';
@@ -113,10 +157,14 @@ class AttachmentTypeController
             return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
         }
 
-        //TODO que faire quand on supprime
-        $attachmentType = AttachmentTypeModel::getById(['select' => 1, 'id' => $args['id']]);
+        $attachmentType = AttachmentTypeModel::getById(['select' => ['type_id'], 'id' => $args['id']]);
         if (empty($attachmentType)) {
             return $response->withStatus(400)->withJson(['errors' => 'Attachment type does not exist']);
+        }
+
+        $attachments = AttachmentModel::get(['select' => 1, 'where' => ['attachment_type = ?', 'status != ?'], 'data' => [$attachmentType['type_id'], 'DEL']]);
+        if (!empty($attachments)) {
+            return $response->withStatus(400)->withJson(['errors' => 'Type is used in attachments']);
         }
 
         AttachmentTypeModel::delete([
